requirements.txt
pandas>=2.2.2
pyarrow>=17.0.0
zstandard>=0.22.0
PyYAML>=6.0.2
tqdm>=4.66.5
numpy>=1.26.4,<2
torch>=2.3.1
scikit-learn
